중간고사 커밋

+++ dongyangSD-master/app/src/main/java/com/example/campmate/data/model/Review.kt
@@ -1,7 +1,7 @@
 package com.example.campmate.data.model
 
 data class Review(
-    val reviewId: Int,
+    val reviewId: Long,
     val campsiteId: Int, // ✅✅✅ [추가] 빠져있던 campsiteId 필드입니다. ✅✅✅
     val campsiteName: String,
     val authorName: String,
